Add TestReporter to __init__ & move testgenreporter + docs
Moved TestGenReporter to it's own file and better documented it's
intended use case with warnings.
