+++ src/testlink/__init__.py
@@ -22,4 +22,4 @@ from .testlinkerrors import TestLinkError
 from .testlinkapigeneric import TestlinkAPIGeneric
 from .testlinkapi import TestlinkAPIClient
 from .testlinkhelper import TestLinkHelper
-from .testreporter import TestGenReporter
+from .testreporter import TestReporter, TestGenReporter
